I want to work on a Dabo Book over the next year or so, so I copied the docbook structure from the SVN Book and got started with a tentative outline. I don't necessarily want to be 100% owner of the book, so if others are interested in writing/editing this as well let's try to get coordinated.
And if it looks like we'll actually have a finished product that is well-written
and printable/sellable, we'll need to find a publisher and pitch the idea. Anyone 
know one of those? <g>

Thank you to the authors of the Subversion Book for allowing me to copy/modify 
your structure so I had something to start with.